sci-service/requirements.txt
python3-git